+++ sdk/identity/azure-identity/src/test/java/com/azure/identity/implementation/msalextensions/PersistentTokenCacheAccessAspectTest.java
@@ -69,7 +69,9 @@ public class PersistentTokenCacheAccessAspectTest {
 
     @After
     public void cleanup() {
-        accessAspect.deleteCache();
+        if (accessAspect != null) {
+            accessAspect.deleteCache();
+        }
     }
 
     @Test
